pyproject.toml
[project]
name = "minitwitter"
version = "0.1.0"
description = "REST API for a simple social media platform"
authors = [
    {name = "Arthur Emanuel Souza Cassiano da Costa",email = "arthurcosta0ac@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
package-mode=false

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
extend-exclude = ['manage.py', '**/migrations/**.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'